Simplify hawk_session/credentials parameter validation
Checking for a falsey value rather than `None` is both more strict (a
good thing) and more readable. This will particularly be the case when
we replace `credentials` with two variables, since we'll need to enforce
"either hawk_session OR both id AND key".

Also removes the redundant `elif` (it's after a `raise`).