Simplify the API for using HawkAuth when the id and key are known
Previously a `credentials` dict would have to have been constructed,
that contained the `id`, `key`, and `algorithm` - that could then be
passed to the `HawkAuth` constructor. However a dict is less
user-friendly than just passing the parameters directly - and the
algorithm rarely varied in practice, so it makes more sense for that to
have a default.

As such, the `HawkAuth` constructor has been changed to accept `id` and
`key` parameters, and the existing `credentials` parameter has been
removed. The algorithm defaults to `sha256`, but can be overridden with
a new `algorithm` parameter. This parameter also allows overriding the
algorithm when deriving the credentials from the hawk session token,
which wasn't previously possible.

Passing the `credentials` parameter now results in an exception with
migration advice - as such the major version of this package must be
bumped for the next release.

Fixes #8.